Use tags for Kolla checkout
With this change we can ensure that we don't unintentionally deploy
patches that have been merged into the cumulus/rocky branch of
kolla-ansible and that we can upgrade Kolla docker images
individually.

The tag format is cumulus/{kolla-release}.{downstream_patch_number}.

Eg. cumulus/7.0.1.1 is the 7.0.1 release of Kolla, with a backported
patch set. If we developed or backported a second patch set to the
7.0.1 release, the tag would change to cumulus/7.0.1.2.