Try to stop fluentd stopping sending to elastic
reload_connections false

The above is chosen as we are behind haproxy, so we should not reload
the list of hosts from elasticsearch.

reconnect_on_error true

Normally connections are not reconnected on error, so eventually the
pool is full, but all the connections are hung. Seems like this happens
after the 10,000 requests that trigger the above reload, or maybe when
haproxy fails over to a new host. Its not totally clear.